DEPRECATED/ISSUE/requirements.txt
python-dotenv>=1.0,<2.0
selenium>=4.20,<5.0
webdriver-manager>=4.0,<5.0
